Java code - Junit 5 tests included and written out
This code simply takes the arguments number1, number2, and number3, it checks for the basic math functions +, -, *, /, and returns the value as a string. any combination of number1 and number2 will equal number 3, the checks are for what method of math function is used to reach that number. However if number1 and number 2 do not equal anything the program will return "none". 

Included are JUnit5 test methods for basic arithmetic calculations that can be changed at any time.